python-tools/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
avro==1.11.3
fastavro==1.8.2
requests==2.31.0
python-dotenv==1.0.0